fix: extract getStreamContext to fix build error
